requirements.txt
streamlit==1.16.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.76
pillow==9.4.0
numpy==1.24.2
